core/pyproject.toml
[tool.poetry]
name = "medalflow"
version = "2.0.0"
description = "MedalFlow: metadata-driven, multi-cloud ETL orchestration with auto-generated DAGs"
authors = ["nadeem4.nk13@gmail.com"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/nadeem4/medalflow.git"
repository = "https://github.com/nadeem4/medalflow.git"
keywords = ["etl", "metadata", "multi-cloud", "synapse", "databricks", "fabric", "auto-dag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
packages = [{include = "core", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
pandas = "^2.0.0"
sqlalchemy = "^2.0.0"
azure-identity = "^1.15.0"
azure-storage-file-datalake = "^12.14.0"
opentelemetry-api = "^1.22.0"
sqlglot = "^23.0.0"
pyodbc = {version = "^5.0.0", optional = true}

[tool.poetry.group.dev.dependencies]
# Essential testing tools
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
# Code quality essentials
black = "^23.12.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
# Simplified to essential rules only
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = ["E501"]  # Line too long - handled by black

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/core"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
